% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{The work horse of the package}
\usage{
mcmc(N, m, no_of_cores, dir = ".")
}
\arguments{
\item{N}{The number of design points to attempt per m}

\item{m}{A vector (see paper)}

\item{no_of_cores}{Integer specifying the number of cores.}

\item{dir}{Directory to save the output file}
}
\description{
The work horse of the package
}
\details{
Since a single aphid run takes less than a single, we actually generate
4 * no_of_cores threads. This is a bit more efficient.
}
